+++ frontend/src/components/Home/Home.module.scss
@@ -1,4 +1,4 @@
-@import "../../../assets/styles/colors.module.scss";
+@import "../../assets/styles/colors.module.scss";
 
 body {
   background-color: $principal-blue;
